Deriv A.P.I Trading bot/requirements.txt
websockets==10.4
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
dash==2.11.1
plotly==5.15.0
talib-binary==0.4.24
scipy==1.11.1
gunicorn==21.2.0 